+++ packages/shared-src/src/reports/covid-vaccine-daily-summary-village.js
@@ -3,16 +3,12 @@ import { differenceInYears, subDays } from 'date-fns';
 import config from 'config';
 import moment from 'moment';
 import { keyBy } from 'lodash';
-import { DATA_TIME_FORMAT, TupaiaApi } from './TupaiaApiStub';
+import { DATA_TIME_FORMAT } from '@tupaia/api-client';
 import { generateReportFromQueryData } from './utilities';
 
 const reportColumnTemplate = [
-  {
-    title: 'village',
-    accessor: data => data.village,
-  },
-  { title: 'tupaiaEntityCode', accessor: data => data.tupaiaEntityCode },
-  { title: 'data_time', accessor: data => data.data_time },
+  { title: 'entity_code', accessor: data => data.tupaiaEntityCode },
+  { title: 'timestamp', accessor: data => data.data_time },
   { title: 'COVIDVac1', accessor: data => data.COVIDVac1 },
   { title: 'COVIDVac2', accessor: data => data.COVIDVac2 },
   { title: 'COVIDVac3', accessor: data => data.COVIDVac3 },
@@ -199,18 +195,29 @@ function addTupaiaEntityCodes(data, villages) {
   }));
 }
 
-async function getVillages() {
-  const tupaiaApi = new TupaiaApi();
+async function getVillages(tupaiaApi) {
+  const reportConfig = config.reports?.['covid-vaccine-daily-summary-village'];
 
-  const countryName = config.country?.name;
+  if (!reportConfig) {
+    throw new Error('Report not configured');
+  }
 
-  return tupaiaApi.getEntities(countryName, 'village');
+  const { hierarchyName, countryCode } = reportConfig;
+
+  const entities = await tupaiaApi.entity.getDescendantsOfEntity(hierarchyName, countryCode, {
+    fields: ['code', 'name', 'type'],
+    filter: {
+      type: 'village',
+    },
+  });
+
+  return entities;
 }
 
-export async function dataGenerator(models, parameters) {
+export async function dataGenerator(models, parameters, tupaiaApi) {
   const listData = await queryCovidVaccineListData(models, parameters);
 
-  const villages = await getVillages();
+  const villages = await getVillages(tupaiaApi);
 
   const tupaiaListData = addTupaiaEntityCodes(listData, villages);
 
@@ -220,3 +227,5 @@ export async function dataGenerator(models, parameters) {
 }
 
 export const permission = 'PatientVaccine';
+
+export const needsTupaiaApiClient = true;